docs: keep size review in CI
