+++ scripts/check-affected/model.ts
@@ -59,7 +59,6 @@ export type CheckId =
   | 'tmpdir-leaks'
   | 'tmpdir-leaks-model'
   | 'coverage-model'
-  | 'pr-evidence-model'
   | 'wire-compat-model'
   | 'production-exports'
   | 'bundle-owner-files'
@@ -119,7 +118,6 @@ export const ALL_CHECKS: readonly CheckId[] = [
   'tmpdir-leaks',
   'tmpdir-leaks-model',
   'coverage-model',
-  'pr-evidence-model',
   'wire-compat-model',
   'production-exports',
   'bundle-owner-files',
